kata: refactoring Factory Method project. Moving types to its own files, adjusting scope access; Renaming project for better identification;
