+++ ConsoleApp1/Page.cs
@@ -1,0 +1,6 @@
+﻿namespace ConApp_Exercises_2;
+
+/// <summary>
+/// The 'Product' abstract class
+/// </summary>
+abstract class Page {  }
