+++ ConsoleApp1/Report.cs
@@ -1,0 +1,9 @@
+﻿namespace ConApp_Exercises_2;
+
+internal class Report : Document
+{
+	public override void CreatePages()
+	{
+		Pages.Add(new SkillsPage());
+	}
+}
