+++ ConsoleApp1/Resume.cs
@@ -1,0 +1,9 @@
+﻿namespace ConApp_Exercises_2;
+
+internal class Resume : Document
+{
+	public override void CreatePages()
+	{
+		Pages.Add(new SkillsPage());
+	}
+}
