+++ ConsoleApp1/SkillsPage.cs
@@ -1,0 +1,14 @@
+﻿using static System.Console;
+
+namespace ConApp_Exercises_2;
+
+/// <summary>
+/// A 'ConcreteProduct' class
+/// </summary>
+class SkillsPage : Page
+{
+	public static void WhatAmIm()
+	{
+		WriteLine($"I am SkillsPage");
+	}
+}
